Changed discord message structure to json, add unique time to delete messages

+++ bungeecord/src/main/java/de/staticred/discordbot/discordcommands/HelpCommandExecutor.java
@@ -1,5 +1,7 @@
 package de.staticred.discordbot.discordcommands;
 
+import de.staticred.discordbot.files.ConfigFileManager;
+import de.staticred.discordbot.files.DiscordMessageFileManager;
 import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.entities.Member;
 import net.dv8tion.jda.api.entities.Message;
@@ -21,15 +23,30 @@ public class HelpCommandExecutor {
         if(args.length != 1) {
             embedBuilder.setDescription("Please use '!help'");
             embedBuilder.setColor(Color.RED);
-            tc.sendMessage(embedBuilder.build()).queue(msg -> msg.delete().queueAfter(10, TimeUnit.SECONDS));
-            command.delete().queueAfter(10,TimeUnit.SECONDS);
+
+            int time = ConfigFileManager.INSTANCE.getTime();
+
+            if(time != -1) {
+                tc.sendMessage(embedBuilder.build()).queue(msg -> msg.delete().queueAfter(time, TimeUnit.SECONDS));
+                command.delete().queueAfter(10,TimeUnit.SECONDS);
+            }else{
+                tc.sendMessage(embedBuilder.build()).queue();
+            }
+
             return;
         }
 
         embedBuilder.setDescription("!help -> list of all commands\n!verify -> synchronise your discord rank with your minecraft rank\n!unlink -> unlink from your mc account\n!update -> update your discord rank");
         embedBuilder.setColor(Color.GREEN);
-        tc.sendMessage(embedBuilder.build()).queue(msg -> msg.delete().queueAfter(10, TimeUnit.SECONDS));
-        command.delete().queueAfter(10,TimeUnit.SECONDS);
+
+        int time = ConfigFileManager.INSTANCE.getTime();
+
+        if(time != -1) {
+            tc.sendMessage(embedBuilder.build()).queue(msg -> msg.delete().queueAfter(time, TimeUnit.SECONDS));
+            command.delete().queueAfter(10,TimeUnit.SECONDS);
+        }else{
+            tc.sendMessage(embedBuilder.build()).queue();
+        }
 
     }
